Modification Design + Barre de Recherche

+++ Modeles/accesBD.php
@@ -273,7 +273,7 @@ class accesBD
 		//-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 
 
-		public function supprimerEmprunt($idClient,$idSupport)
+		public function supprimerEmprunt($idClient,$idSupport,$action=false)
 		{
 			$today = date("Y-m-d");
 			//Verification Presence Emprunt ou Non
@@ -281,18 +281,19 @@ class accesBD
 			$requete->execute();
 
 			if($requete->fetch()[0]!='0') {
+				if($action) {
+					//---------Suppresion d'un Emprunt--------------
+					$requete = $this->conn->prepare("DELETE FROM emprunt WHERE idSupport = (?) AND idClient=(?) AND dateFinEmprunt >= (?) ;");
+					$requete->bindValue(1,$idSupport);
+					$requete->bindValue(2,$idClient);
+					$requete->bindValue(3,$today);
 
-				//---------Suppresion d'un Emprunt--------------
-				$requete = $this->conn->prepare("DELETE FROM emprunt WHERE idSupport = (?) AND idClient=(?) AND dateFinEmprunt >= (?) ;");
-				$requete->bindValue(1,$idSupport);
-				$requete->bindValue(2,$idClient);
-				$requete->bindValue(3,$today);
+					if(!$requete->execute())
+					{
+						die("Erreur dans SupprimerEmprunt : ".$requete->errorCode());
+					}
 
-				if(!$requete->execute())
-				{
-					die("Erreur dans SupprimerEmprunt : ".$requete->errorCode());
 				}
-
 				return true;
 			}
 
@@ -376,18 +377,22 @@ class accesBD
 
 
 		//Donne Image Séries
-		public function donneImageSerie()
+		public function donneImageSerie($idGenre)
 		{
 			$liste = '';
-			$requete = $this->conn->prepare("SELECT image,s.idSupport FROM SUPPORT s, SERIE se WHERE se.idSupport = s.idSupport GROUP BY se.idSerie");
+			$requete = $this->conn->prepare("SELECT image,s.idSupport FROM SUPPORT s, SERIE se WHERE se.idSupport = s.idSupport AND s.idGenre LIKE ? GROUP BY se.idSerie");
+			$requete->bindValue(1,$idGenre.'%');
+
 			$requete->execute();
 			return $requete;
 		}
 
-		public function donneImageFilm()
+		public function donneImageFilm($idGenre)
 		{
 			$liste = '';
-			$requete = $this->conn->prepare("SELECT image,s.idSupport FROM SUPPORT s, FILM f WHERE f.idSupport = s.idSupport GROUP BY f.idFilm");
+			$requete = $this->conn->prepare("SELECT image,s.idSupport FROM SUPPORT s, FILM f WHERE f.idSupport = s.idSupport AND s.idGenre LIKE ? GROUP BY f.idFilm");
+			$requete->bindValue(1,$idGenre.'%');
+
 			$requete->execute();
 			return $requete;
 		}
@@ -404,6 +409,16 @@ class accesBD
 			return $requete;
 		}
 
+		public function donneImagesGenre($idGenre='%')
+		{
+
+			$requete = $this->conn->prepare("SELECT s.image,s.idGenre,s.idSupport FROM support s WHERE s.idGenre LIKE ? GROUP BY s.idGenre");
+			$requete->bindValue(1,$idGenre);
+
+			$requete->execute();
+
+			return $requete;
+		}
 
 		//-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 		//----------------------------- RETOURNER L IMAGE VIDEO - RESEARCH BAR ---------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -411,11 +426,27 @@ class accesBD
 
 
 		//PROCEDURE pour recuperer les informations de la vidéo mis en barre de recherche
-		public function retournerInfosRecherche($video)
+		public function retournerInfosRecherche($video,$typeSupport)
 		{
-				$requete = $this->conn->prepare("SELECT idSupport,titreSupport,image FROM support WHERE titreSupport LIKE ?	LIMIT 10;");
+			//Recherche en fonction du Type du Support = > Serie / Film / Emprunt / ...
+				switch($typeSupport)
+				{
+					case 'serie':
+						$requete = $this->conn->prepare("SELECT s.idSupport,titreSupport,image FROM support s,serie se WHERE se.idSupport = s.idSupport AND titreSupport LIKE ?	LIMIT 10;");
+						$requete->bindValue(1,$video.'%');
+						break;
 
-				$requete->bindValue(1,$video.'%');
+					case 'film':
+						$requete = $this->conn->prepare("SELECT s.idSupport,titreSupport,image FROM support s,film f WHERE f.idSupport = s.idSupport AND titreSupport LIKE ?	LIMIT 10;");
+						$requete->bindValue(1,$video.'%');
+						break;
+					default:
+							$requete = $this->conn->prepare("SELECT idSupport,titreSupport,image FROM support WHERE titreSupport LIKE ?	LIMIT 10;");
+							$requete->bindValue(1,$video.'%');
+							break;
+
+				}
+
 				$requete->execute();
 
 				$requete = $requete->fetchAll();
@@ -425,35 +456,36 @@ class accesBD
 		}
 
 		//-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-		//----------------------------- RETOURNER INFOS SUPPORT - Click Support ---------------------------------------------------------------------------------------------------------------------------------------------------------------
+		//----------------------------- RETOURNER EMPRUNTER OU NON ---------------------------------------------------------------------------------------------------------------------------------------------------------------
 		//-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-		/*public function retournerInfos($id,$type)
+		/*
+		public function emprunter($idSupport)
 		{
 				//Si ça nous retourne des infos d'un film alors on retourne cette requête
-				if($type='Film')
-				$requete = $this->retournerInfosFilm($id);
-				var_dump(mysql_num_rows($requete));
-				if($requete)
-				{
-					return $requete;
-				}
+				$today = date("Y-m-d");
+				$requete = $this->conn->prepare("SELECT COUNT(*) FROM SUPPORT s WHERE idSupport = (?) AND dateFinEmprunt >= (?)");
 
-				$requete = $this->retournerInfosSerie($id);
+				$requete->bindValue(1,$idSupport);
+				$requete->bindValue(2,$today);
+
+				$requete->execute();
 
 				if($requete)
 				{
-					return $requete;
+					if($requete->fetch()[0]!="0")
+					{
+						return true;
+					}
 				}
 
-				return null;
+				return false;
 
 		}*/
 
 
 		public function retournerInfosFilm($idSupport)
 		{
-			$requete = $this->conn->prepare("SELECT titreSupport,realisateur,libelleGenre,duree FROM support,genre,film WHERE film.idSupport=support.idSupport AND support.idGenre = genre.idGenre AND support.idSupport = ? LIMIT 1");
+			$requete = $this->conn->prepare("SELECT image,titreSupport,realisateur,libelleGenre,duree FROM support,genre,film WHERE film.idSupport=support.idSupport AND support.idGenre = genre.idGenre AND support.idSupport = ? LIMIT 1");
 
 			$requete->bindValue(1,$idSupport);
 			$requete->execute();
@@ -469,7 +501,7 @@ class accesBD
 
 		public function retournerInfosSerie($idSupport)
 		{
-			$requete = $this->conn->prepare("SELECT titreSupport,realisateur,libelleGenre,resumeSerie FROM support,genre,serie WHERE serie.idSupport=support.idSupport AND support.idGenre = genre.idGenre AND support.idSupport = ? LIMIT 1");
+			$requete = $this->conn->prepare("SELECT image,titreSupport,realisateur,libelleGenre,resumeSerie FROM support,genre,serie WHERE serie.idSupport=support.idSupport AND support.idGenre = genre.idGenre AND support.idSupport = ? LIMIT 1");
 
 			$requete->bindValue(1,$idSupport);
 			$requete->execute();
